fix up of writer tests
