--- StockGrader/StockGrader.Test/TestFiles/ARK_CHANGED_REDUCED_TESLA.csv
@@ -1,30 +1,0 @@
-date,fund,company,ticker,cusip,shares,"market value ($)","weight (%)"
-04/03/2023,ARKK,"TESLA INC",TSLA,88160R101,"3,986,021","$826,939,916.66",8%
-04/03/2023,ARKK,"ZOOM VIDEO COMMUNICATIONS-A",ZM,98980L101,"8,709,980","$643,144,923.20",8.07%
-04/03/2023,ARKK,"ROKU INC",ROKU,77543R102,"9,409,480","$619,331,973.60",7.77%
-04/03/2023,ARKK,"COINBASE GLOBAL INC -CLASS A",COIN,19260Q107,"7,871,089","$531,849,483.73",6.67%
-04/03/2023,ARKK,"EXACT SCIENCES CORP",EXAS,30063P105,"7,266,556","$492,745,162.36",6.18%
-04/03/2023,ARKK,"BLOCK INC",SQ,852234103,"7,049,389","$483,940,554.85",6.07%
-04/03/2023,ARKK,"UIPATH INC - CLASS A",PATH,90364P105,"27,220,325","$477,988,907.00",6.00%
-04/03/2023,ARKK,"SHOPIFY INC - CLASS A",SHOP,82509L107,"9,254,741","$443,672,283.54",5.57%
-04/03/2023,ARKK,"TELADOC HEALTH INC",TDOC,87918A105,"12,353,461","$319,954,639.90",4.01%
-04/03/2023,ARKK,"DRAFTKINGS INC-CL A","DKNG UW",26142V105,"16,226,013","$314,135,611.68",3.94%
-04/03/2023,ARKK,"ROBLOX CORP -CLASS A",RBLX,771049103,"6,578,183","$295,886,671.34",3.71%
-04/03/2023,ARKK,"UNITY SOFTWARE INC",U,91332U101,"8,801,815","$285,530,878.60",3.58%
-04/03/2023,ARKK,"TWILIO INC - A",TWLO,90138F102,"4,224,971","$281,509,817.73",3.53%
-04/03/2023,ARKK,"PAGERDUTY INC",PD,69553P100,"7,607,000","$266,092,860.00",3.34%
-04/03/2023,ARKK,"INTELLIA THERAPEUTICS INC",NTLA,45826J105,"6,900,786","$257,192,294.22",3.23%
-04/03/2023,ARKK,"CRISPR THERAPEUTICS AG",CRSP,H17182108,"5,613,888","$253,916,154.24",3.19%
-04/03/2023,ARKK,"ROBINHOOD MARKETS INC - A",HOOD,770700102,"20,215,037","$196,288,009.27",2.46%
-04/03/2023,ARKK,"BEAM THERAPEUTICS INC",BEAM,07373V105,"6,235,329","$190,925,773.98",2.40%
-04/03/2023,ARKK,"PACIFIC BIOSCIENCES OF CALIF",PACB,69404D108,"14,730,466","$170,578,796.28",2.14%
-04/03/2023,ARKK,"10X GENOMICS INC-CLASS A",TXG,88025U109,"2,936,812","$163,844,741.48",2.06%
-04/03/2023,ARKK,"GINKGO BIOWORKS HOLDINGS INC",DNA,37611X100,"106,641,901","$141,833,728.33",1.78%
-04/03/2023,ARKK,"VERACYTE INC",VCYT,92337F107,"4,869,365","$108,586,839.50",1.36%
-04/03/2023,ARKK,"TWIST BIOSCIENCE CORP",TWST,90184D100,"3,804,918","$57,378,163.44",0.72%
-04/03/2023,ARKK,"2U INC",TWOU,90214J101,"6,447,115","$44,162,737.75",0.55%
-04/03/2023,ARKK,"CERUS CORP",CERS,157085101,"14,356,040","$42,637,438.80",0.53%
-04/03/2023,ARKK,"INVITAE CORP",NVTA,46185L103,"18,580,194","$25,083,261.90",0.31%
-04/03/2023,ARKK,"VERVE THERAPEUTICS INC",VERV,92539P101,"1,570,508","$22,646,725.36",0.28%
-04/03/2023,ARKK,"GOLDMAN FS TRSY OBLIG INST 468",,X9USDGSFT,"13,637,423","$13,637,422.70",0.17%
-"The principal risks of investing in ARK ETFs include equity, market, management and non-diversification risks, as well as fluctuations in market value and net asset value ('NAV'). Investors should carefully consider the investment objectives and risks as well as charges and expenses of an ETF before investing. This and other information are contained in each ETF's prospectus, which may be obtained by contacting the ETF's transfer agent The Bank of New York Mellon or by clicking here. Please read the prospectus carefully before you invest. The market price of ETF shares may differ significantly from their NAV during periods of market volatility. ETF shares may only be redeemed directly with the ETF at NAV by Authorized Participants, in very large creation units. Holdings are subject to change without notice and are not a recommendation to buy or sell any security. There can be no guarantee that an active trading market for ETF shares will develop or be maintained, or that their listing will continue or remain unchanged. Buying or selling ETF shares on an exchange may require the payment of brokerage commissions and frequent trading may incur brokerage costs that detract significantly from investment returns. Copyright 2023, ARK Investment Management LLC. All content is original and has been researched and produced by ARK Investment Management LLC unless otherwise stated. No part of this content may be reproduced in any form, or referred to in any other publication, without the express written permission of ARK Investment Management LLC. All statements made regarding companies, securities or other financial information on this site or any sites or articles relating to ARK Investment Management LLC are strictly beliefs and points of view held by ARK Investment Management and are subject to change without notice. Information contained herein has been obtained from sources believed to be reliable, but not guaranteed. This material has been distributed for informational purposes only and should not be considered as investment advice or a recommendation of any particular security, strategy or investment product."
