+++ StockGrader/StockGrader.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33402.96
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StockGrader.Writer.Test", "StockGrader.Test\StockGrader.Writer.Test.csproj", "{D01FB383-6B87-4B20-8CE6-C854A692BE0A}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StockGrader.ExecutableConsoleApp", "StockGrader.ExecutableConsoleApp\StockGrader.ExecutableConsoleApp.csproj", "{470311CE-744F-49AE-A4C8-6D3F5029BAAD}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StockGrader.DAL", "StockGrader.DAL\StockGrader.DAL.csproj", "{4BFBAE53-E0E7-419D-9E2D-A42ECCAAD81D}"
@@ -13,7 +11,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StockGrader.BL", "StockGrad
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StockGrader.Runner", "StockGrader.Runner\StockGrader.Runner.csproj", "{F7B081D7-FE5E-473A-8522-A6E7F25BE9D8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StockGrader.BL.Test", "StockGrader.BL.Test\StockGrader.BL.Test.csproj", "{E2618532-F254-4D4E-AD27-59D8ED85028B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StockGrader.BL.Test", "StockGrader.BL.Test\StockGrader.BL.Test.csproj", "{E2618532-F254-4D4E-AD27-59D8ED85028B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,10 +19,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D01FB383-6B87-4B20-8CE6-C854A692BE0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D01FB383-6B87-4B20-8CE6-C854A692BE0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D01FB383-6B87-4B20-8CE6-C854A692BE0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D01FB383-6B87-4B20-8CE6-C854A692BE0A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{470311CE-744F-49AE-A4C8-6D3F5029BAAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{470311CE-744F-49AE-A4C8-6D3F5029BAAD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{470311CE-744F-49AE-A4C8-6D3F5029BAAD}.Release|Any CPU.ActiveCfg = Release|Any CPU
